Fix #171, start on #174 model example updates
