+++ release-notes/0.4.6.md
@@ -1,4 +1,5 @@
 # Release notes for sbmlutils 0.4.6
 - update to libsbml>=5.19.0
 - update libroadrunner>=2.0.4
+- bugfix #173, deprecated xlrd replaced with openpyxl
