+++ src/sbmlutils/examples/models/amount_species/model.py
@@ -1,4 +1,3 @@
-# -*- coding=utf-8 -*-
 """
 Model with amount species
 """
