+++ src/sbmlutils/examples/models/annotation/results/annotation_example_8.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:fbc="http://www.sbml.org/sbml/level3/version1/fbc/version2" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" level="3" version="1" fbc:required="false" comp:required="true">
-  <model metaid="meta_annotation_example_8" id="annotation_example_8" name="annotation_example_8" substanceUnits="mole" timeUnits="s" volumeUnits="m3" areaUnits="m2" lengthUnits="m" extentUnits="mole" fbc:strict="false">
+<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" level="3" version="1" comp:required="true">
+  <model metaid="meta_annotation_example_8" id="annotation_example_8" name="annotation_example_8" substanceUnits="mole" timeUnits="s" volumeUnits="m3" areaUnits="m2" lengthUnits="m" extentUnits="mole">
     <notes>
       <body xmlns="http://www.w3.org/1999/xhtml">
         <h1>Model with inline annotations</h1>
@@ -11,7 +11,7 @@
         <div class="dc:publisher">This file has been created by
     <a href="https://livermetabolism.com" title="Matthias Koenig" target="_blank">Matthias Koenig</a>.</div>
         <h2>Terms of use</h2>
-        <div class="dc:rightsHolder">Copyright © 2019 Matthias Koenig.</div>
+        <div class="dc:rightsHolder">Copyright © 2021 Matthias Koenig.</div>
         <div class="dc:license">
           <p>Redistribution and use of any part of this model, with or without modification, are permitted provided
         that the following conditions are met:
@@ -46,10 +46,10 @@
             </rdf:Bag>
           </dcterms:creator>
           <dcterms:created rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-06-23T21:23:39Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:created>
           <dcterms:modified rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-06-23T21:23:39Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:modified>
         </rdf:Description>
       </rdf:RDF>
@@ -250,5 +250,14 @@
         </kineticLaw>
       </reaction>
     </listOfReactions>
+    <comp:listOfPorts>
+      <comp:port metaid="kg_port" sboTerm="SBO:0000599" comp:unitRef="kg" comp:id="kg_port" comp:name="kg_port"/>
+      <comp:port metaid="s_port" sboTerm="SBO:0000599" comp:unitRef="s" comp:id="s_port" comp:name="s_port"/>
+      <comp:port metaid="m_port" sboTerm="SBO:0000599" comp:unitRef="m" comp:id="m_port" comp:name="m_port"/>
+      <comp:port metaid="m2_port" sboTerm="SBO:0000599" comp:unitRef="m2" comp:id="m2_port" comp:name="m2_port"/>
+      <comp:port metaid="m3_port" sboTerm="SBO:0000599" comp:unitRef="m3" comp:id="m3_port" comp:name="m3_port"/>
+      <comp:port metaid="mM_port" sboTerm="SBO:0000599" comp:unitRef="mM" comp:id="mM_port" comp:name="mM_port"/>
+      <comp:port metaid="mole_per_s_port" sboTerm="SBO:0000599" comp:unitRef="mole_per_s" comp:id="mole_per_s_port" comp:name="mole_per_s_port"/>
+    </comp:listOfPorts>
   </model>
 </sbml>
