+++ src/sbmlutils/examples/models/assignment/assignment_example.py
@@ -2,14 +2,16 @@
 """
 Simple assignment test case.
 """
+from pathlib import Path
+
 from sbmlutils.factory import *
 from sbmlutils.modelcreator import templates
+from sbmlutils.modelcreator.creator import create_model
 from sbmlutils.units import *
 
 
 # ---------------------------------------------------------------------------------------------------------------------
-mid = "assignment"
-version = 4
+mid = "assignment_example"
 creators = templates.creators
 notes = Notes(
     [
@@ -74,21 +76,6 @@ units = [
     ),
 ]
 
-# ---------------------------------------------------------------------------------------------------------------------
-# --- Functions ---
-# ---------------------------------------------------------------------------------------------------------------------
-
-# ---------------------------------------------------------------------------------------------------------------------
-# Compartments
-# ---------------------------------------------------------------------------------------------------------------------
-
-# ---------------------------------------------------------------------------------------------------------------------
-# Species
-# ---------------------------------------------------------------------------------------------------------------------
-
-# ---------------------------------------------------------------------------------------------------------------------
-# Parameters
-# ---------------------------------------------------------------------------------------------------------------------
 parameters = [
     # dosing
     Parameter("Ave", 0, "mg", constant=False),
@@ -101,17 +88,11 @@ parameters = [
     Parameter("FVve", 0.0514, "litre_per_kg", True),
 ]
 
-# ---------------------------------------------------------------------------------------------------------------------
-# Assignments
-# ---------------------------------------------------------------------------------------------------------------------
 assignments = [
     InitialAssignment("Ave", "IVDOSE", "mg"),
     InitialAssignment("D", "PODOSE", "mg"),
 ]
 
-# ---------------------------------------------------------------------------------------------------------------------
-# Rules
-# ---------------------------------------------------------------------------------------------------------------------
 rules = [
     # concentrations
     AssignmentRule("Cve", "Ave/Vve", "mg_per_litre"),
@@ -123,8 +104,16 @@ rate_rules = [
     RateRule("Ave", "- k1*Cve", "mg_per_h"),
 ]
 
-# ---------------------------------------------------------------------------------------------------------------------
-# Reactions
-# ---------------------------------------------------------------------------------------------------------------------
-
 reactions = []
+
+
+def create(tmp=False):
+    create_model(
+        modules=["sbmlutils.examples.models.assignment.assignment_example"],
+        output_dir=Path(__file__).parent / "results",
+        tmp=tmp,
+    )
+
+
+if __name__ == "__main__":
+    create()
