+++ src/sbmlutils/examples/models/demo/model.py
@@ -11,6 +11,7 @@ from . import reactions as R
 
 # ---------------------------------------------------------------------------------------------------------------------
 mid = "Koenig_demo"
+packages = ["fbc"]
 version = 14
 notes = Notes(
     [
