+++ src/sbmlutils/examples/models/demo/results/Koenig_demo_14_no_annotations.html
@@ -446,7 +446,7 @@
         <!--------------------------------------------------->
         <a class="anchor" id="anchor_model"></a>
         <h1 id="model"><a href="Koenig_demo_14_no_annotations.xml" target="_blank" alt="Access SBML model" title="SBML model"><i class="fas fa-file-code"></i></a>&nbsp;Model : Koenig_demo_14</h1>
-        <b>Creator</b><br />Matthias, Koenig, Humboldt-University Berlin, Institute for Theoretical Biology, <a href="mailto:koenigmx@hu-berlin.de" target="_blank">koenigmx@hu-berlin.de</a><br /><b>Created:</b> 2019-11-12 21:39<br /><b>Modified:</b> 2019-11-12 21:39<br /><br />
+        <b>Creator</b><br />Matthias, Koenig, Humboldt-University Berlin, Institute for Theoretical Biology, <a href="mailto:koenigmx@hu-berlin.de" target="_blank">koenigmx@hu-berlin.de</a><br /><b>Created:</b> 1900-01-01 00:00<br /><b>Modified:</b> 1900-01-01 00:00<br /><br />
         <table id="table_model" class="table table-condensed table-nonfluid">
             <tr>
             <td>
@@ -464,7 +464,7 @@
                     </table>
                 </td>
                 <td>
-                    <span class="package">L3V1</span> <span class="package">fbc-V2</span> <span class="package">comp-V1</span><br />
+                    <span class="package">L3V1</span> <span class="package">comp-V1</span> <span class="package">fbc-V2</span><br />
                     <div class="cvterm"></div><br />
                     <div class="cvterm"></div>
                 </td>
@@ -488,7 +488,7 @@
     <div class="dc:publisher">This file has been created by
     <a href="https://livermetabolism.com" title="Matthias Koenig" target="_blank">Matthias Koenig</a>.</div>
     <h2>Terms of use</h2>
-    <div class="dc:rightsHolder">Copyright © 2019 Matthias Koenig.</div>
+    <div class="dc:rightsHolder">Copyright © 2021 Matthias Koenig.</div>
     <div class="dc:license">
       <p>Redistribution and use of any part of this model, with or without modification, are permitted provided
         that the following conditions are met:
