+++ src/sbmlutils/examples/models/distrib/results/distrib_comp_example.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" level="3" version="1" comp:required="true">
-  <model metaid="meta_initial_assignment" id="initial_assignment" name="initial_assignment" substanceUnits="mmole" timeUnits="min" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mmole">
+<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" xmlns:distrib="http://www.sbml.org/sbml/level3/version1/distrib/version1" level="3" version="1" comp:required="true" distrib:required="true">
+  <model metaid="meta_distrib_comp_example" id="distrib_comp_example" name="distrib_comp_example" substanceUnits="mole" timeUnits="hr" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mole">
     <notes>
       <body xmlns="http://www.w3.org/1999/xhtml">
-        <p>Example model for testing InitialAssignments in roadrunner.</p>
+        <h1>sbmlutils {}</h1>
+        <h2>Description</h2>
+        <p>Example creating distrib model with distribution elements.</p>
         <div class="dc:provenance">The content of this model has been carefully created in a manual research effort.</div>
         <div class="dc:publisher">This file has been created by
     <a href="https://livermetabolism.com" title="Matthias Koenig" target="_blank">Matthias Koenig</a>.</div>
@@ -27,7 +29,7 @@
     </notes>
     <annotation>
       <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:vCard="http://www.w3.org/2001/vcard-rdf/3.0#" xmlns:vCard4="http://www.w3.org/2006/vcard/ns#" xmlns:bqbiol="http://biomodels.net/biology-qualifiers/" xmlns:bqmodel="http://biomodels.net/model-qualifiers/">
-        <rdf:Description rdf:about="#meta_initial_assignment">
+        <rdf:Description rdf:about="#meta_distrib_comp_example">
           <dcterms:creator>
             <rdf:Bag>
               <rdf:li rdf:parseType="Resource">
@@ -52,14 +54,9 @@
       </rdf:RDF>
     </annotation>
     <listOfUnitDefinitions>
-      <unitDefinition id="min">
+      <unitDefinition id="hr">
         <listOfUnits>
-          <unit kind="second" exponent="1" scale="0" multiplier="60"/>
-        </listOfUnits>
-      </unitDefinition>
-      <unitDefinition id="mmole">
-        <listOfUnits>
-          <unit kind="mole" exponent="1" scale="-3" multiplier="1"/>
+          <unit kind="second" exponent="1" scale="0" multiplier="3600"/>
         </listOfUnits>
       </unitDefinition>
       <unitDefinition id="m">
@@ -73,31 +70,31 @@
         </listOfUnits>
       </unitDefinition>
     </listOfUnitDefinitions>
-    <listOfCompartments>
-      <compartment id="c" spatialDimensions="3" size="2" units="litre" constant="true"/>
-    </listOfCompartments>
-    <listOfSpecies>
-      <species id="A1" compartment="c" initialAmount="1" substanceUnits="mmole" hasOnlySubstanceUnits="true" boundaryCondition="false" constant="false"/>
-    </listOfSpecies>
     <listOfParameters>
-      <parameter id="D" value="5" units="mmole" constant="true"/>
+      <parameter id="p1" value="1" units="mole" constant="true">
+        <distrib:listOfUncertainties>
+          <distrib:uncertainty>
+            <distrib:uncertParameter distrib:value="2" distrib:type="mean"/>
+            <distrib:uncertParameter distrib:value="2" distrib:type="standardDeviation"/>
+            <distrib:uncertSpan distrib:type="range" distrib:valueLower="1" distrib:valueUpper="4"/>
+            <distrib:uncertParameter distrib:type="distribution" distrib:definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal">
+              <math xmlns="http://www.w3.org/1998/Math/MathML">
+                <apply>
+                  <csymbol encoding="text" definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal"> normal </csymbol>
+                  <cn> 2 </cn>
+                  <cn> 2 </cn>
+                </apply>
+              </math>
+            </distrib:uncertParameter>
+          </distrib:uncertainty>
+        </distrib:listOfUncertainties>
+      </parameter>
     </listOfParameters>
-    <listOfInitialAssignments>
-      <initialAssignment symbol="A1">
-        <math xmlns="http://www.w3.org/1998/Math/MathML" xmlns:sbml="http://www.sbml.org/sbml/level3/version1/core">
-          <apply>
-            <times/>
-            <ci> D </ci>
-            <cn sbml:units="dimensionless" type="integer"> 2 </cn>
-          </apply>
-        </math>
-      </initialAssignment>
-    </listOfInitialAssignments>
     <comp:listOfPorts>
-      <comp:port metaid="min_port" sboTerm="SBO:0000599" comp:unitRef="min" comp:id="min_port" comp:name="min_port"/>
-      <comp:port metaid="mmole_port" sboTerm="SBO:0000599" comp:unitRef="mmole" comp:id="mmole_port" comp:name="mmole_port"/>
+      <comp:port metaid="hr_port" sboTerm="SBO:0000599" comp:unitRef="hr" comp:id="hr_port" comp:name="hr_port"/>
       <comp:port metaid="m_port" sboTerm="SBO:0000599" comp:unitRef="m" comp:id="m_port" comp:name="m_port"/>
       <comp:port metaid="m2_port" sboTerm="SBO:0000599" comp:unitRef="m2" comp:id="m2_port" comp:name="m2_port"/>
+      <comp:port metaid="p1_port" sboTerm="SBO:0000599" comp:idRef="p1" comp:id="p1_port" comp:name="p1_port"/>
     </comp:listOfPorts>
   </model>
 </sbml>
