+++ src/sbmlutils/examples/models/distrib/results/distrib_comp_example_flat.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:fbc="http://www.sbml.org/sbml/level3/version1/fbc/version2" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" level="3" version="1" fbc:required="false" comp:required="true">
-  <model metaid="meta_example1_1" id="example1_1" name="example1_1" substanceUnits="mmole" timeUnits="min" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mmole" fbc:strict="false">
+<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:distrib="http://www.sbml.org/sbml/level3/version1/distrib/version1" level="3" version="1" distrib:required="true">
+  <model metaid="meta_distrib_comp_example" id="distrib_comp_example" name="distrib_comp_example" substanceUnits="mole" timeUnits="hr" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mole">
     <notes>
       <body xmlns="http://www.w3.org/1999/xhtml">
         <h1>sbmlutils {}</h1>
         <h2>Description</h2>
-        <p>Example model showing how to create compartments, species and reactions.</p>
+        <p>Example creating distrib model with distribution elements.</p>
         <div class="dc:provenance">The content of this model has been carefully created in a manual research effort.</div>
         <div class="dc:publisher">This file has been created by
     <a href="https://livermetabolism.com" title="Matthias Koenig" target="_blank">Matthias Koenig</a>.</div>
         <h2>Terms of use</h2>
-        <div class="dc:rightsHolder">Copyright © 2019 Matthias Koenig.</div>
+        <div class="dc:rightsHolder">Copyright © 2021 Matthias Koenig.</div>
         <div class="dc:license">
           <p>Redistribution and use of any part of this model, with or without modification, are permitted provided
         that the following conditions are met:
@@ -29,7 +29,7 @@
     </notes>
     <annotation>
       <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:vCard="http://www.w3.org/2001/vcard-rdf/3.0#" xmlns:vCard4="http://www.w3.org/2006/vcard/ns#" xmlns:bqbiol="http://biomodels.net/biology-qualifiers/" xmlns:bqmodel="http://biomodels.net/model-qualifiers/">
-        <rdf:Description rdf:about="#meta_example1_1">
+        <rdf:Description rdf:about="#meta_distrib_comp_example">
           <dcterms:creator>
             <rdf:Bag>
               <rdf:li rdf:parseType="Resource">
@@ -45,15 +45,20 @@
             </rdf:Bag>
           </dcterms:creator>
           <dcterms:created rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-03-17T18:47:27Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:created>
           <dcterms:modified rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-03-17T18:47:27Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:modified>
         </rdf:Description>
       </rdf:RDF>
     </annotation>
     <listOfUnitDefinitions>
+      <unitDefinition id="hr">
+        <listOfUnits>
+          <unit kind="second" exponent="1" scale="0" multiplier="3600"/>
+        </listOfUnits>
+      </unitDefinition>
       <unitDefinition id="m">
         <listOfUnits>
           <unit kind="metre" exponent="1" scale="0" multiplier="1"/>
@@ -64,45 +69,26 @@
           <unit kind="metre" exponent="2" scale="0" multiplier="1"/>
         </listOfUnits>
       </unitDefinition>
-      <unitDefinition id="min">
-        <listOfUnits>
-          <unit kind="second" exponent="1" scale="0" multiplier="60"/>
-        </listOfUnits>
-      </unitDefinition>
-      <unitDefinition id="mmole">
-        <listOfUnits>
-          <unit kind="mole" exponent="1" scale="-3" multiplier="1"/>
-        </listOfUnits>
-      </unitDefinition>
-      <unitDefinition id="mM">
-        <listOfUnits>
-          <unit kind="mole" exponent="1" scale="-3" multiplier="1"/>
-          <unit kind="litre" exponent="-1" scale="0" multiplier="1"/>
-        </listOfUnits>
-      </unitDefinition>
-      <unitDefinition id="mole_per_min">
-        <listOfUnits>
-          <unit kind="mole" exponent="1" scale="0" multiplier="1"/>
-          <unit kind="second" exponent="-1" scale="0" multiplier="60"/>
-        </listOfUnits>
-      </unitDefinition>
     </listOfUnitDefinitions>
-    <listOfCompartments>
-      <compartment id="c" spatialDimensions="3" size="1" units="litre" constant="true"/>
-    </listOfCompartments>
-    <listOfSpecies>
-      <species sboTerm="SBO:0000247" id="S1" name="S1" compartment="c" initialConcentration="5" substanceUnits="mmole" hasOnlySubstanceUnits="false" boundaryCondition="false" constant="false"/>
-    </listOfSpecies>
     <listOfParameters>
-      <parameter sboTerm="SBO:0000371" id="R1_Km" name="Km R1" value="0.1" units="mM" constant="true"/>
-      <parameter sboTerm="SBO:0000186" id="R1_Vmax" name="Vmax R1" value="10" units="mole_per_min" constant="true"/>
+      <parameter id="p1" value="1" units="mole" constant="true">
+        <distrib:listOfUncertainties>
+          <distrib:uncertainty>
+            <distrib:uncertParameter distrib:value="2" distrib:type="mean"/>
+            <distrib:uncertParameter distrib:value="2" distrib:type="standardDeviation"/>
+            <distrib:uncertSpan distrib:type="range" distrib:valueLower="1" distrib:valueUpper="4"/>
+            <distrib:uncertParameter distrib:type="distribution" distrib:definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal">
+              <math xmlns="http://www.w3.org/1998/Math/MathML">
+                <apply>
+                  <csymbol encoding="text" definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal"> normal </csymbol>
+                  <cn> 2 </cn>
+                  <cn> 2 </cn>
+                </apply>
+              </math>
+            </distrib:uncertParameter>
+          </distrib:uncertainty>
+        </distrib:listOfUncertainties>
+      </parameter>
     </listOfParameters>
-    <listOfInitialAssignments>
-      <initialAssignment symbol="S1">
-        <math xmlns="http://www.w3.org/1998/Math/MathML" xmlns:sbml="http://www.sbml.org/sbml/level3/version1/core">
-          <cn sbml:units="mM"> 10 </cn>
-        </math>
-      </initialAssignment>
-    </listOfInitialAssignments>
   </model>
 </sbml>
