+++ src/sbmlutils/examples/models/distrib/results/uncertainty_example.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:fbc="http://www.sbml.org/sbml/level3/version1/fbc/version2" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" xmlns:distrib="http://www.sbml.org/sbml/level3/version1/distrib/version1" level="3" version="1" fbc:required="false" comp:required="true" distrib:required="true">
-  <model metaid="meta_uncertainty_ex1_1" id="uncertainty_ex1_1" name="uncertainty_ex1_1" substanceUnits="mole" timeUnits="h" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mole" fbc:strict="false">
+<sbml xmlns="http://www.sbml.org/sbml/level3/version1/core" xmlns:comp="http://www.sbml.org/sbml/level3/version1/comp/version1" xmlns:distrib="http://www.sbml.org/sbml/level3/version1/distrib/version1" level="3" version="1" comp:required="true" distrib:required="true">
+  <model metaid="meta_uncertainty_example" id="uncertainty_example" name="uncertainty_example" substanceUnits="mole" timeUnits="hr" volumeUnits="litre" areaUnits="m2" lengthUnits="m" extentUnits="mole">
     <notes>
       <body xmlns="http://www.w3.org/1999/xhtml">
         <h1>sbmlutils {}</h1>
@@ -10,7 +10,7 @@
         <div class="dc:publisher">This file has been created by
     <a href="https://livermetabolism.com" title="Matthias Koenig" target="_blank">Matthias Koenig</a>.</div>
         <h2>Terms of use</h2>
-        <div class="dc:rightsHolder">Copyright © 2019 Matthias Koenig.</div>
+        <div class="dc:rightsHolder">Copyright © 2021 Matthias Koenig.</div>
         <div class="dc:license">
           <p>Redistribution and use of any part of this model, with or without modification, are permitted provided
         that the following conditions are met:
@@ -29,7 +29,7 @@
     </notes>
     <annotation>
       <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:vCard="http://www.w3.org/2001/vcard-rdf/3.0#" xmlns:vCard4="http://www.w3.org/2006/vcard/ns#" xmlns:bqbiol="http://biomodels.net/biology-qualifiers/" xmlns:bqmodel="http://biomodels.net/model-qualifiers/">
-        <rdf:Description rdf:about="#meta_uncertainty_ex1_1">
+        <rdf:Description rdf:about="#meta_uncertainty_example">
           <dcterms:creator>
             <rdf:Bag>
               <rdf:li rdf:parseType="Resource">
@@ -45,16 +45,16 @@
             </rdf:Bag>
           </dcterms:creator>
           <dcterms:created rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-03-29T00:25:30Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:created>
           <dcterms:modified rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-03-29T00:25:30Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>1900-01-01T00:00:00Z</dcterms:W3CDTF>
           </dcterms:modified>
         </rdf:Description>
       </rdf:RDF>
     </annotation>
     <listOfUnitDefinitions>
-      <unitDefinition id="h">
+      <unitDefinition id="hr">
         <listOfUnits>
           <unit kind="second" exponent="1" scale="0" multiplier="3600"/>
         </listOfUnits>
@@ -77,10 +77,23 @@
             <distrib:uncertParameter distrib:value="2" distrib:type="mean"/>
             <distrib:uncertParameter distrib:value="2" distrib:type="standardDeviation"/>
             <distrib:uncertSpan distrib:type="range" distrib:valueLower="1" distrib:valueUpper="4"/>
-            <distrib:uncertParameter distrib:type="distribution" distrib:definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal"/>
+            <distrib:uncertParameter distrib:type="distribution" distrib:definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal">
+              <math xmlns="http://www.w3.org/1998/Math/MathML">
+                <apply>
+                  <csymbol encoding="text" definitionURL="http://www.sbml.org/sbml/symbols/distrib/normal"> normal </csymbol>
+                  <cn> 2 </cn>
+                  <cn> 2 </cn>
+                </apply>
+              </math>
+            </distrib:uncertParameter>
           </distrib:uncertainty>
         </distrib:listOfUncertainties>
       </parameter>
     </listOfParameters>
+    <comp:listOfPorts>
+      <comp:port metaid="hr_port" sboTerm="SBO:0000599" comp:unitRef="hr" comp:id="hr_port" comp:name="hr_port"/>
+      <comp:port metaid="m_port" sboTerm="SBO:0000599" comp:unitRef="m" comp:id="m_port" comp:name="m_port"/>
+      <comp:port metaid="m2_port" sboTerm="SBO:0000599" comp:unitRef="m2" comp:id="m2_port" comp:name="m2_port"/>
+    </comp:listOfPorts>
   </model>
 </sbml>
