+++ src/sbmlutils/modelcreator/creator.py
@@ -15,13 +15,14 @@ import logging
 import shutil
 import tempfile
 from pathlib import Path
-from typing import Dict, Iterable, List
+from typing import Dict, Iterable, List, NamedTuple
 
 import libsbml
 import xmltodict
 
 import sbmlutils.factory as factory
 import sbmlutils.history as history
+from build.lib.sbmlutils.modelcreator.creator import CoreModel
 from sbmlutils.factory import SBML_LEVEL, SBML_VERSION
 from sbmlutils.io import write_sbml
 from sbmlutils.metadata import annotator
@@ -33,55 +34,16 @@ from sbmlutils.validation import check
 logger = logging.getLogger(__name__)
 
 
-class Factory:
-    """
-    Generic model factory, which should be subclassed by the individual
-    ModelFactories.
-    """
-
-    def __init__(
-        self, modules: Iterable[str], output_dir: Path, annotations=None, mid=None
-    ):
-        """
-
-        :param modules: iterable of module strings; should be importable as is
-        :param output_dir: path to output directory
-        :param annotations:
-        :param mid:
-        """
-        self.modules = modules
-        self.target_dir = output_dir
-        self.annotations = annotations
-        self.mid = mid
-
-    def create(self, tmp=False):
-        """Creates SBML model in target directory.
-
-        :param tmp: write files in temporary folder. Used for testing.
-        :return: (model_dict, core_model, sbml_path)
-        """
-        if tmp:
-            output_dir = tempfile.mkdtemp()
-        else:
-            output_dir = self.target_dir
-
-        try:
-            [model_dict, core_model, sbml_path] = create_model(
-                modules=self.modules,
-                output_dir=output_dir,
-                annotations=self.annotations,
-                mid=self.mid,
-            )
-        finally:
-            if tmp:
-                shutil.rmtree(output_dir)
-
-        return [model_dict, core_model, sbml_path]
+class FactoryResult(NamedTuple):
+    model_dict: Dict
+    core_model: CoreModel
+    sbml_path: Path
 
 
 def create_model(
-    modules: List[str],
+    modules: Iterable[str],
     output_dir: Path,
+    tmp: bool = False,
     filename: str = None,
     mid: str = None,
     suffix: str = None,
@@ -92,7 +54,7 @@ def create_model(
     units_consistency: bool = True,
     modeling_practice: bool = True,
     internal_consistency: bool = True,
-):
+) -> FactoryResult:
     """Create SBML model from module information.
 
     This is the entry point for creating models.
@@ -100,15 +62,21 @@ def create_model(
     If no filename is provided the filename is created from the id and suffix.
     Additional model annotations can be provided.
 
-    :param modules: iteratable of strings of python modules
+    :param modules: iterable of strings of python modules
     :param output_dir: directory in which to create SBML file
+    :param tmp: boolean flag to create files in a temporary directory (for testing)
     :param filename: filename to write to with suffix, if not provided mid and suffix are used
     :param mid: model id to use for filename
     :param suffix: suffix for SBML filename
     :param annotations: list of annotations for SBML
     :param create_report: boolean switch to create SBML report
     :param validate: validates the SBML file
-    :return:
+    :param log_errors: boolean flag to log errors
+    :param units_consistency: boolean flag to check units consistency
+    :param modeling_practice: boolean flag to check modeling practise
+    :param internal_consistency: boolean flag to check internal consistency
+
+    :return: FactoryResult
     """
     # preprocess
     logger.info(
@@ -129,14 +97,16 @@ def create_model(
     logger.debug(core_model.get_info())
     core_model.create_sbml()
 
-    # write file
-    if isinstance(output_dir, str):
-        output_dir = Path(output_dir)
-        logger.warning(f"'output_dir' should be a Path: {output_dir}")
+    if tmp:
+        output_dir = tempfile.mkdtemp()
+    else:
+        if isinstance(output_dir, str):
+            output_dir = Path(output_dir)
+            logger.warning(f"'output_dir' should be a Path: {output_dir}")
 
-    if not output_dir.exists():
-        logger.warning(f"'output_dir' does not exist and is created: {output_dir}")
-        output_dir.mkdir(parents=True)
+        if not output_dir.exists():
+            logger.warning(f"'output_dir' does not exist and is created: {output_dir}")
+            output_dir.mkdir(parents=True)
 
     if not filename:
         # create filename
@@ -150,31 +120,39 @@ def create_model(
     sbml_path = output_dir / filename
     if core_model.doc is None:
         core_model.create_sbml()
-    write_sbml(
-        doc=core_model.doc,
-        filepath=sbml_path,
-        validate=validate,
-        log_errors=log_errors,
-        units_consistency=units_consistency,
-        modeling_practice=modeling_practice,
-        internal_consistency=internal_consistency,
+    try:
+        write_sbml(
+            doc=core_model.doc,
+            filepath=sbml_path,
+            validate=validate,
+            log_errors=log_errors,
+            units_consistency=units_consistency,
+            modeling_practice=modeling_practice,
+            internal_consistency=internal_consistency,
+        )
+
+        # annotate
+        if annotations is not None:
+            # overwrite the normal file
+            annotator.annotate_sbml(
+                source=sbml_path, annotations_path=annotations, filepath=sbml_path
+            )
+
+        # create report
+        if create_report:
+            # file is already validated, no validation on report needed
+            sbmlreport.create_report(
+                sbml_path=sbml_path, output_dir=output_dir, validate=False
+            )
+    finally:
+        if tmp:
+            shutil.rmtree(output_dir)
+
+    return FactoryResult(
+        model_dict=model_dict,
+        core_model=core_model,
+        sbml_path=sbml_path,
     )
-
-    # annotate
-    if annotations is not None:
-        # overwrite the normal file
-        annotator.annotate_sbml(
-            source=sbml_path, annotations_path=annotations, filepath=sbml_path
-        )
-
-    # create report
-    if create_report:
-        # file is already validated, no validation on report needed
-        sbmlreport.create_report(
-            sbml_path=sbml_path, output_dir=output_dir, validate=False
-        )
-
-    return [model_dict, core_model, sbml_path]
 
 
 class Preprocess:
