+++ src/sbmlutils/test/modelcreator/test_modelcreator.py
@@ -6,12 +6,11 @@ and creation of the SBML.
 """
 import pytest
 
-from sbmlutils.examples.models.annotation import factory as annotation_factory
-from sbmlutils.examples.models.assignment import factory as assignment_factory
-from sbmlutils.examples.models.basic import factory as basic_factory
+from sbmlutils.examples.models.annotation import annotation_example
+from sbmlutils.examples.models.assignment import assignment_example
+from sbmlutils.examples.models.basic import basic_example1, basic_example2
 from sbmlutils.examples.models.demo import factory as demo_factory
-from sbmlutils.examples.models.distrib import factory as distrib_factory
-from sbmlutils.examples.models.example1 import factory as example1_factory
+from sbmlutils.examples.models.distrib import distributions_example, distrib_comp_example, uncertainty_example
 from sbmlutils.examples.models.fbc import factory as fbc_factory
 from sbmlutils.examples.models.mass_charge_balance import factory as mass_charge_factory
 from sbmlutils.examples.models.tiny_model import factory as tiny_factory
@@ -23,24 +22,22 @@ def test_create_mass_charge():
 
 
 def test_create_annotation():
-    """Create assignment model.
-    :return:
-    """
-    annotation_factory.create(tmp=True)
+    annotation_example.create(tmp=True)
+
+
+def test_create_distrib():
+    distrib_comp_example.create(tmp=True)
+    distributions_example.create(tmp=True)
+    uncertainty_example.create(tmp=True)
 
 
 def test_create_assignment():
-    """Create assignment model.
-    :return:
-    """
-    assignment_factory.create(tmp=True)
+    assignment_example.create(tmp=True)
 
 
 def test_create_basic():
-    """Create basic model.
-    :return:
-    """
-    basic_factory.create(tmp=True)
+    basic_example1.create(tmp=True)
+    basic_example2.create(tmp=True)
 
 
 def test_create_demo():
