+++ src/sbmlutils/test/test_distrib.py
@@ -8,6 +8,21 @@ from sbmlutils.modelcreator.creator import CoreModel
 from sbmlutils.units import *
 from sbmlutils.validation import validate_doc
 
+from sbmlutils.examples.models.distrib import distrib_comp_example, \
+    distributions_example, uncertainty_example
+
+
+def test_distrib_comp_example():
+    distrib_comp_example.create(tmp=True)
+
+
+def test_distributions_example():
+    distributions_example.create(tmp=True)
+
+
+def test_uncertainty_example():
+    uncertainty_example.create(tmp=True)
+
 
 def check_model_dict(d: Dict) -> libsbml.SBMLDocument:
     """Check that no errors."""
